requirements.txt
Flask
mysql-connector-python
pandas
openpyxl
werkzeug